changes after PR comments
